+++ tomcat/src/test/java/com/techcourse/RequestMappingTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package com.techcourse;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,10 +6,11 @@ import com.techcourse.controller.LoginController;
 import com.techcourse.controller.RegisterController;
 import java.util.List;
 import java.util.stream.Stream;
+import org.apache.catalina.Controller;
+import org.apache.catalina.resource.ResourceHandler;
 import org.apache.coyote.http11.data.HttpMethod;
 import org.apache.coyote.http11.data.HttpRequest;
 import org.apache.coyote.http11.data.HttpVersion;
-import org.apache.coyote.http11.resource.ResourceHandler;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
